requirements-dev.txt
aniso8601==9.0.1
asgiref==3.7.2
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-filter==22.1
django-graphql-jwt==0.4.0
djangorestframework==3.14.0
Faker==19.3.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
graphene==3.3
graphene-django==3.1.5
graphene-django-extras==1.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
Pillow==10.0.0
promise==2.3
proto-plus==1.22.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
